Implemented Hide Button
After testing the purpose of the played? button, it was quickly noticed that because of the nature of the combinations it would only make it LESS likely to play someone you haven't played yet (e.g., by the 4th or 5th match players may encounter a repeated opponent). By the time it is easy to predict (3-4 players remaining) the player will be able to mentally recognize who they are going to play next  anecdotally. Instead, I created a hide button feature, this will become useful for when players get eliminated from the game and if there is a mistake, they could always unhide the player as well.